add a few small fixes
