+++ src/components/Navigation/Navigation.tsx
@@ -1,0 +1,36 @@
+import classNames from 'classnames';
+import { NavLink } from 'react-router-dom';
+
+export const Navigation = () => {
+  const activeNav
+  = ({ isActive }: { isActive: boolean }) => classNames('navbar-item', {
+    'has-background-grey-lighter': isActive,
+  });
+
+  return (
+    <nav
+      data-cy="nav"
+      className="navbar is-fixed-top has-shadow"
+      role="navigation"
+      aria-label="main navigation"
+    >
+      <div className="container">
+        <div className="navbar-brand">
+          <NavLink
+            className={activeNav}
+            to="/"
+          >
+            Home
+          </NavLink>
+
+          <NavLink
+            className={activeNav}
+            to="/people"
+          >
+            People
+          </NavLink>
+        </div>
+      </div>
+    </nav>
+  );
+};
