--- src/components/PeoplePage/index.tsx
@@ -1,1 +1,0 @@
-export * from './PeoplePage';
