+++ src/components/pages/HomePage.tsx
@@ -1,0 +1,5 @@
+export const HomePage = () => {
+  return (
+    <h1 className="title">Home Page</h1>
+  );
+};
